# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface District {
  id: string;
  nameUz: string;
  nameRu: string;
  nameEn: string;
  lat: number;
  lng: number;
  available: boolean;
  /** @nullable */
  osmRelationId?: string | null;
}

export interface Service {
  id: string;
  nameUz: string;
  nameRu: string;
  nameEn: string;
  price: number;
  extraPickupPrice: number;
  /** @nullable */
  description?: string | null;
}

export interface PublicSettings {
  servicePrice: number;
  pickupExtra: number;
  paymentCard: string;
  paymentOwner: string;
  clickPaymentUrl: string;
  adminContact?: string;
}

export type OrderInputLang = typeof OrderInputLang[keyof typeof OrderInputLang];


export const OrderInputLang = {
  uz: 'uz',
  ru: 'ru',
  en: 'en',
} as const;

export type OrderInputPatientType = typeof OrderInputPatientType[keyof typeof OrderInputPatientType];


export const OrderInputPatientType = {
  adult: 'adult',
  child: 'child',
} as const;

export type OrderInputPatientGender = typeof OrderInputPatientGender[keyof typeof OrderInputPatientGender];


export const OrderInputPatientGender = {
  male: 'male',
  female: 'female',
} as const;

export interface OrderInput {
  telegramUserId: number;
  /** @nullable */
  telegramUsername?: string | null;
  lang: OrderInputLang;
  patientType: OrderInputPatientType;
  patientName: string;
  patientAge: number;
  patientGender: OrderInputPatientGender;
  serviceId: string;
  /** @nullable */
  childTiming?: string | null;
  /** @nullable */
  usesDiaper?: boolean | null;
  complaints: string[];
  /** @nullable */
  customComplaint?: string | null;
  deliverySlot: string;
  /** @nullable */
  pickupSlot?: string | null;
  districtId: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  addressNote?: string | null;
}

export interface Order {
  id: number;
  orderId: string;
  telegramUserId?: number;
  patientName?: string;
  patientAge?: number;
  patientGender?: string;
  patientType?: string;
  serviceId?: string;
  complaints?: string[];
  deliverySlot?: string;
  /** @nullable */
  pickupSlot?: string | null;
  districtId?: string;
  latitude?: number;
  longitude?: number;
  price?: number;
  extraPrice?: number;
  isFree?: boolean;
  status: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}