# lib/api-zod/src/generated/types/district.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */

export interface District {
  id: string;
  nameUz: string;
  nameRu: string;
  nameEn: string;
  lat: number;
  lng: number;
  available: boolean;
  /** @nullable */
  osmRelationId?: string | null;
}