# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  orderId: string;
  telegramUserId?: number;
  patientName?: string;
  patientAge?: number;
  patientGender?: string;
  patientType?: string;
  serviceId?: string;
  complaints?: string[];
  deliverySlot?: string;
  /** @nullable */
  pickupSlot?: string | null;
  districtId?: string;
  latitude?: number;
  longitude?: number;
  price?: number;
  extraPrice?: number;
  isFree?: boolean;
  status: string;
  createdAt: string;
}