# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputLang } from './orderInputLang';
import type { OrderInputPatientGender } from './orderInputPatientGender';
import type { OrderInputPatientType } from './orderInputPatientType';

export interface OrderInput {
  telegramUserId: number;
  /** @nullable */
  telegramUsername?: string | null;
  lang: OrderInputLang;
  patientType: OrderInputPatientType;
  patientName: string;
  patientAge: number;
  patientGender: OrderInputPatientGender;
  serviceId: string;
  /** @nullable */
  childTiming?: string | null;
  /** @nullable */
  usesDiaper?: boolean | null;
  complaints: string[];
  /** @nullable */
  customComplaint?: string | null;
  deliverySlot: string;
  /** @nullable */
  pickupSlot?: string | null;
  districtId: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  addressNote?: string | null;
}