# lib/api-zod/src/generated/types/publicSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicSettings {
  servicePrice: number;
  pickupExtra: number;
  paymentCard: string;
  paymentOwner: string;
  clickPaymentUrl: string;
  adminContact?: string;
}