# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedBot Web App API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  nameUz: string;
  nameRu: string;
  nameEn: string;
  price: number;
  extraPickupPrice: number;
  /** @nullable */
  description?: string | null;
}